@@ -1,12 +1,9 @@
-const Router = require('koa-router');
-const api = new Router();
+import Router from 'koa-router';
+import posts from './posts';
 
-const posts = require('./posts');
+const api = new Router();
 
 api.use('/posts', posts.routes());
 
-api.get('/test', (ctx) => {
-  ctx.body = 'test';
-});
-
-module.exports = api;
+// 라우터를 내보냅니다.
+export default api;
