@@ -1,15 +1,5 @@
-const Koa = require('koa');
-const Router = require('koa-router');
-const bodyParser = require('koa-bodyparser');
+// 이 파일에서만 no-global-assign ESLint 옵션을 비활성화합니다
+/* eslint-disable no-global-assign */
 
-const app = new Koa();
-const router = new Router();
-
-const api = require('./api');
-
-router.use('/api', api.routes());
-app.use(bodyParser());
-app.use(router.routes()).use(router.allowedMethods());
-app.listen(4000, () => {
-  console.log('Listening to port 4000');
-});
+require = require('esm')(module /*, options*/);
+module.exports = require('./main.js');
